# api-service/internal/items/http/app/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.GetFollowersResponse:
    properties:
      followers:
        items:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.User'
        type: array
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.GetFollowingResponse:
    properties:
      following:
        items:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.User'
        type: array
    type: object
...
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.User:
    properties:
      bio:
        type: string
      created_at:
        $ref: '#/definitions/timestamppb.Timestamp'
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      profile_picture_url:
        type: string
      updated_at:
        $ref: '#/definitions/timestamppb.Timestamp'
      username:
        type: string
    type: object
...
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. Negative
          second values with fractions must still have non-negative nanos values
          that count forward in time. Must be from 0 to 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch
          1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
          9999-12-31T23:59:59Z inclusive.
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Endpoints for MiniTwitter
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: '# MiniTwitter'
  version: 1.03.67.83.145
paths:
...
  /user/follow/{user_id}:
    post:
      consumes:
      - application/json
      description: Follow a user by their user ID
      parameters:
      - description: User ID to follow
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - Users
  /user/followers:
    get:
      consumes:
      - application/json
      description: Get the list of followers for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.GetFollowersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get followers
      tags:
      - Users
  /user/following:
    get:
      consumes:
      - application/json
      description: Get the list of users the authenticated user is following
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.GetFollowingResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get following
      tags:
      - Users
...
  /user/unfollow/{user_id}:
    post:
      consumes:
      - application/json
      description: Unfollow a user by their user ID
      parameters:
      - description: User ID to unfollow
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - Users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"